pyproject.toml
[tool.poetry]
name = "hades"
version = "0.1.0"
description = ""
authors = ["Patarimi <38954040+Patarimi@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.scripts]
hades = "hades.main:app"

[tool.poetry.dependencies]
python = "^3.10"
poethepoet = "^0.20.0"
typer = "^0.9.0"
gdstk = "^0.9.42"
setuptools = "^68.0.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
